# docs/swagger.yaml
definitions:
...
  responses.FailureBody:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  responses.SuccessBody:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  schemas.UpdateUserSchema:
    properties:
      date_of_birth:
        description: Keep as string
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      mobile_number:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Fetches user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.FailureBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.FailureBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.FailureBody'
      security:
      - AuthorizationToken: []
      summary: Get User Details
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Updates user details by user ID. Only the provided fields will
        be updated.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateUserSchema'
      produces:
      - application/json
      responses:
        "200":
          description: User details updated successfully
          schema:
            $ref: '#/definitions/responses.SuccessBody'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/responses.FailureBody'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/responses.FailureBody'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.FailureBody'
      security:
      - AuthorizationToken: []
      summary: Update User Details
      tags:
      - User
securityDefinitions:
  AuthorizationToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"